backend/requirements.txt
fastapi
uvicorn
pymongo
requests
schedule
python-multipart
python-socketio
numpy
scipy
matplotlib
pygame
pyqtree
tqdm